backend/app/auth/jwt.py
from datetime import UTC, datetime, timedelta
from uuid import UUID

from jose import JWTError, jwt

from app.config import settings

ALGORITHM = "HS256"


def create_access_token(user_id: UUID) -> str:
    expire = datetime.now(UTC) + timedelta(minutes=settings.access_token_expire_minutes)
    payload = {
        "sub": str(user_id),
        "exp": expire,
        "type": "access",
    }
    return jwt.encode(payload, settings.secret_key, algorithm=ALGORITHM)


def create_refresh_token(user_id: UUID) -> str:
    expire = datetime.now(UTC) + timedelta(days=settings.refresh_token_expire_days)
    payload = {
        "sub": str(user_id),
        "exp": expire,
        "type": "refresh",
    }
    return jwt.encode(payload, settings.secret_key, algorithm=ALGORITHM)


def decode_token(token: str) -> dict:
    try:
        return jwt.decode(token, settings.secret_key, algorithms=[ALGORITHM])
    except JWTError as e:
        raise ValueError(str(e)) from e


def verify_access_token(token: str) -> UUID:
    payload = decode_token(token)
    if payload.get("type") != "access":
        raise ValueError("Not an access token")
    user_id = payload.get("sub")
    if user_id is None:
        raise ValueError("Token missing subject")
    return UUID(user_id)


def verify_refresh_token(token: str) -> UUID:
    payload = decode_token(token)
    if payload.get("type") != "refresh":
        raise ValueError("Not a refresh token")
    user_id = payload.get("sub")
    if user_id is None:
        raise ValueError("Token missing subject")
    return UUID(user_id)
